+++ iOSDFULibrary/Classes/Implementation/SecureDFU/Services/SecureDFUService.swift
@@ -47,6 +47,16 @@ import CoreBluetooth
     private let service                       : CBService
     private var dfuPacketCharacteristic       : SecureDFUPacket?
     private var dfuControlPointCharacteristic : SecureDFUControlPoint?
+    
+    /// This method returns true if DFU Control Point characteristc has been discovered.
+    /// A device without this characteristic is not supported and can only be disconnected.
+    internal func supportsReset() -> Bool {
+        // The Abort (0x0C) command has been added to DFU bootloader in SDK 15.
+        // https://infocenter.nordicsemi.com/topic/com.nordic.infocenter.sdk5.v15.0.0/lib_dfu_transport.html?cp=8_5_3_3_5_2
+        // For earlier SDKs there is no way to reset the bootloader other than
+        // disconnecting and waiting for it to time out after few minutes.
+        return dfuControlPointCharacteristic != nil
+    }
 
     private var paused  = false
     private var aborted = false
@@ -55,7 +65,7 @@ import CoreBluetooth
     private var success          : Callback?
     /// A temporary callback used to report an operation error.
     private var report           : ErrorCallback?
-    /// A temporaty callback used to report progress status.
+    /// A temporary callback used to report progress status.
     private var progressDelegate : DFUProgressDelegate?
     private var progressQueue    : DispatchQueue?
     
@@ -188,15 +198,15 @@ import CoreBluetooth
     }
     
     /**
-     Reads Command Object Info. Result it reported using callbacks.
+     Selects the Command Object. Result it reported using callbacks.
      
      - parameter response: Method called when the response was received.
      - parameter report:   Method called when an error occurred.
      */
-    func readCommandObjectInfo(onReponse response: @escaping SecureDFUResponseCallback,
-                               onError report: @escaping ErrorCallback) {
+    func selectCommandObject(onReponse response: @escaping SecureDFUResponseCallback,
+                             onError report: @escaping ErrorCallback) {
         if !aborted {
-            dfuControlPointCharacteristic?.send(.readCommandObjectInfo,
+            dfuControlPointCharacteristic?.send(.selectCommandObject,
                                                 onResponse: response, onError: report)
         } else {
             sendReset(onError: report)
@@ -204,15 +214,15 @@ import CoreBluetooth
     }
     
     /**
-     Reads object info Data. Result it reported using callbacks.
+     Selects the Data Object. Result it reported using callbacks.
      
      - parameter response: Method called when the response was received.
      - parameter report:   Method called when an error occurred.
      */
-    func readDataObjectInfo(onReponse response: @escaping SecureDFUResponseCallback,
-                            onError report: @escaping ErrorCallback) {
+    func selectDataObject(onReponse response: @escaping SecureDFUResponseCallback,
+                          onError report: @escaping ErrorCallback) {
         if !aborted {
-            dfuControlPointCharacteristic?.send(.readDataObjectInfo,
+            dfuControlPointCharacteristic?.send(.selectDataObject,
                                                 onResponse: response, onError: report)
         } else {
             sendReset(onError: report)
@@ -323,10 +333,19 @@ import CoreBluetooth
      
      - parameter report: A callback called when writing characteristic failed.
      */
-    private func sendReset(onError report: @escaping ErrorCallback) {
+    func sendReset(onError report: @escaping ErrorCallback) {
         aborted = true
-        // There is no command to reset a Secure DFU device. We can just disconnect.
-        targetPeripheral?.disconnect()
+        // Upon sending the Abort request the device will immediately reboot in application
+        // mode. There will be no notification with status success returned.
+        dfuControlPointCharacteristic?.send(.abort,
+            onSuccess: nil, // Device will disconnected immediately.
+            onError: { [weak self] _, _ in
+                // Seems like the Abort request is not supported (indicating SDK 12-14).
+                // We can just disconnect. The bootloader should reset to app mode after
+                // a timeout.
+                self?.targetPeripheral?.disconnect()
+            }
+        )
     }
     
     //MARK: - Packet commands